refactor: move etxtbsy
